[노은수] refactor : 로컬 스토리지 검증하는 커스텀 훅으로 분리
